Deep Learning/TransformerUNET Model/requirements.txt
torch>=1.10
torchvision
numpy
tqdm
Pillow
scipy 